rewired EF to make more sense.

+++ CharacterSheet.Infrastructure/Data/Services/ProficiencyService.cs
@@ -19,10 +19,10 @@ namespace CharacterSheet.Infrastructure
         public ClassProficiencyViewModel GetProficienciesByClassId(int classId)
         {
             var clsProficiencies = _characterClassRepository.GetClassProficiencies(classId);
-            var armorProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == ProficiencyTypeId.Armor).ToList();
-            var weaponProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == ProficiencyTypeId.Weapon).ToList();
-            var toolProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == ProficiencyTypeId.Tool).ToList();
-            var saveProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == ProficiencyTypeId.Save).ToList();
+            var armorProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == (int)ProficiencyTypeId.Armor).ToList();
+            var weaponProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == (int)ProficiencyTypeId.Weapon).ToList();
+            var toolProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == (int)ProficiencyTypeId.Tool).ToList();
+            var saveProficiencies = clsProficiencies.Where(p => p.ProficiencytypeId == (int)ProficiencyTypeId.Save).ToList();
 
             return new ClassProficiencyViewModel
             {
